fix: 필드 타입 변경

+++ src/main/java/S13P31A306/loglens/domain/dashboard/service/impl/TopFrequentErrorsQueryServiceImpl.java
@@ -174,7 +174,7 @@ public class TopFrequentErrorsQueryServiceImpl implements TopFrequentErrorsQuery
 
         for (var bucket : buckets) {
             String exceptionType = bucket.key();
-            Long count = bucket.docCount();
+            Integer count = (int) bucket.docCount();
 
             // first/last occurrence
             LocalDateTime firstOccurrence = parseTimestamp(
